+++ backend/pkg/settings/config_test.go
@@ -3,44 +3,10 @@ package settings
 import (
 	"os"
 	"path/filepath"
-	"reflect"
 	"testing"
 
 	"github.com/google/go-cmp/cmp"
 )
-
-func TestInitialize(t *testing.T) {
-	type args struct {
-		configFile string
-	}
-	tests := []struct {
-		name string
-		args args
-	}{
-		// TODO: Add test cases.
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			Initialize(tt.args.configFile)
-		})
-	}
-}
-
-func Test_setDefaults(t *testing.T) {
-	tests := []struct {
-		name string
-		want Settings
-	}{
-		// TODO: Add test cases.
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := SetDefaults(true); !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("setDefaults() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
 
 func TestConfigLoadChanged(t *testing.T) {
 	// Create isolated test directory